Distribución adecuada de almacenamiento

+++ src/services/storage_service.ts
@@ -1,20 +1,27 @@
-import { storageClient, bucketName } from '../config/storage_config';
+// src/services/storage_service.ts
+import { storageClient } from '../config/storage_config';
 import path from 'path';
 import fs from 'fs';
-
+import { BucketService } from './bucket_service';
 
 export class StorageService {
-
-  //Funcion para subir imagen
+  // Función para subir imagen
   static async uploadImage(filePath: string, destinationName?: string) {
     try {
+      // Consultar el bucket con más espacio disponible
+      const bucketService = new BucketService();
+      const bucket = await bucketService.getBucketWithMostAvailableSpace();
+      if (!bucket) {
+        throw new Error('No se encontró un bucket disponible.');
+      }
+      const bucketName = bucket.name;
       const destName = destinationName || path.basename(filePath);
       
       const [file] = await storageClient.bucket(bucketName).upload(filePath, {
         destination: destName,
         metadata: {
-          cacheControl: 'public, max-age=31536000',
-        },
+          cacheControl: 'public, max-age=31536000'
+        }
       });
 
       // Eliminar el archivo local después de subir
@@ -29,9 +36,16 @@ export class StorageService {
     }
   }
 
-  //Funcion para eliminar imagen
+  // Función para eliminar imagen
   static async deleteImage(fileName: string) {
     try {
+      // Consultar el bucket con más espacio disponible
+      const bucketService = new BucketService();
+      const bucket = await bucketService.getBucketWithMostAvailableSpace();
+      if (!bucket) {
+        throw new Error('No se encontró un bucket disponible.');
+      }
+      const bucketName = bucket.name;
       await storageClient.bucket(bucketName).file(fileName).delete();
       return true;
     } catch (error) {
